requirements.txt
motor==2.0.0
discord.py==1.3.3
python-dotenv==0.13.0